fix warnings

+++ agrolib/interpolation/interpolation.cpp
@@ -1769,7 +1769,7 @@ void macroAreaDetrending(Crit3DMacroArea myArea, meteoVariable myVar, Crit3DInte
 
     //find the fitting functions vector based on the length of the parameters vector for every proxy
     std::vector<std::function<double (double, std::vector<double> &)> > fittingFunction;
-    for (int l = 0; l < myArea.getParameters().size(); l++)
+    for (int l = 0; l < (int)myArea.getParameters().size(); l++)
     {
         if (myArea.getParameters()[l].size() == 2)
             fittingFunction.push_back(functionLinear_intercept);
@@ -1785,9 +1785,9 @@ void macroAreaDetrending(Crit3DMacroArea myArea, meteoVariable myVar, Crit3DInte
 
     // create vector of macro area interpolation points
     std::vector<int> temp = myArea.getMeteoPoints();
-    for (int l = 0; l < temp.size(); l++)
+    for (int l = 0; l < (int)temp.size(); l++)
     {
-        for (int k = 0; k < interpolationPoints.size(); k++)
+        for (int k = 0; k < (int)interpolationPoints.size(); k++)
         {
             if (interpolationPoints[k].index == temp[l])
             {
